+++ src/metadata/ffmpeg_thumbnailer_handler.h
@@ -72,6 +72,19 @@ private:
     /// @brief read cached thumbnail
     std::optional<std::vector<std::byte>> readThumbnailCacheFile(const fs::path& movieFilename) const;
 
+    /// @brief size of thumbnails
+    int thumbSize;
+    /// @brief percentage of video to seek for thumbnail
+    int seekPercentage;
+    /// @brief image quality of thumbnails
+    int imageQuality;
+    /// @brief use thumbnail cache
+    bool cacheEnabled;
+    /// @brief add film strip overlay
+    bool stripOverlay;
+    /// @brief rotate images automatically based on orientation
+    bool doRotate;
+
     /// @brief cache generated thumb
     void writeThumbnailCacheFile(
         const fs::path& movieFilename,